+++ tests/unit/test_download.py
@@ -45,7 +45,7 @@ def test_recreate_docs(
         path_file := repository_path / DOCUMENTATION_FOLDER_NAME / "page-path-1" / "page-file-1.md"
     ).is_file()
     assert index_file.read_text(encoding="utf-8") == (
-        f"{index_content}\n\n"
+        f"{index_content}\n"
         "# Contents\n\n"
         "1. [empty-navlink](page-path-1)\n"
         "  1. [file-navlink](page-path-1/page-file-1.md)"